requirements.txt
cm2py
pillow
opencv2-python